requirements.txt
asgiref==3.7.2
attrs==20.3.0
blinker==1.6.2
cachelib==0.10.2
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.6
colorama==0.4.6
cryptography==41.0.3
cs50==9.2.5
Django==5.0.6
django-cors-headers==4.6.0
django-jazzmin==3.0.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.8
Flask==2.3.2
Flask-Mail==0.9.1
Flask-Session==0.5.0
greenlet==3.1.1
idna==3.4
inflection==0.5.1
itsdangerous==2.1.2
jellyfish==0.11.2
Jinja2==3.1.2
jwt==1.3.1
lib50==3.0.7
markdown2==2.4.13
MarkupSafe==2.1.3
numpy==1.26.4
packaging==24.1
pexpect==4.8.0
ptyprocess==0.7.0
pycparser==2.21
PyJWT==2.9.0
pytz==2024.2
PyYAML==6.0.1
requests==2.31.0
rest-framework-simplejwt==0.0.2
setuptools==75.1.0
SQLAlchemy==1.4.46
sqlparse==0.4.4
submit50==3.1.1
termcolor==1.1.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.4
Werkzeug==2.3.7
wheel==0.44.0
